+++ OneByte/Source/Features/Onboarding/View/SubgoalCycleView.swift
@@ -7,7 +7,7 @@ struct SubgoalCycleView: View {
     @Environment(\.modelContext) private var modelContext
     
     @Query private var mainGoals: [MainGoal]
-    @State var viewModel = OnboardingViewModel(createService: CreateService(), updateService: UpdateService(mainGoals: [], subGoals: [], detailGoals: []))
+    @State var viewModel = RoutineCycleViewModel(updateService: UpdateService(mainGoals: [], subGoals: [], detailGoals: []))
     
     @State private var userSubGoal: String = "" // 사용자 SubGoal 입력 텍스트
     @FocusState private var isFocused: Bool // TextField 포커스 상태 관리
